+++ app/src/main/java/com/jin/jjinweather/MainActivity.kt
@@ -16,20 +16,20 @@ import com.jin.jjinweather.feature.location.LocationRepository
 import com.jin.jjinweather.feature.location.data.LocationRepositoryImpl
 import com.jin.jjinweather.feature.locationimpl.data.GeoCodeDataSourceImpl
 import com.jin.jjinweather.feature.locationimpl.data.GeoPointDataSourceImpl
-import com.jin.jjinweather.feature.weather.WeatherRepository
+import com.jin.jjinweather.feature.weather.domain.repository.WeatherRepository
 import com.jin.jjinweather.feature.weather.data.OpenWeatherDataSource
 import com.jin.jjinweather.feature.weather.data.OpenWeatherApi
 import com.jin.jjinweather.feature.weather.data.WeatherRepositoryImpl
 import com.jin.jjinweather.feature.weatherimpl.data.WeatherDataSourceImpl
-import com.jin.jjinweather.layer.data.RetrofitClient
+import com.jin.jjinweather.feature.network.RetrofitClient
 import com.jin.jjinweather.feature.database.data.DatabaseProvider
 import com.jin.jjinweather.feature.datastore.data.PreferencesRepositoryImpl
-import com.jin.jjinweather.layer.domain.usecase.GetCurrentLocationWeatherUseCase
-import com.jin.jjinweather.layer.ui.Screens
-import com.jin.jjinweather.layer.ui.onboarding.OnboardingScreen
-import com.jin.jjinweather.layer.ui.onboarding.OnboardingViewModel
-import com.jin.jjinweather.layer.ui.temperature.TemperatureScreen
-import com.jin.jjinweather.layer.ui.temperature.TemperatureViewModel
+import com.jin.jjinweather.feature.weather.domain.usecase.GetCurrentLocationWeatherUseCase
+import com.jin.jjinweather.feature.navigation.Screens
+import com.jin.jjinweather.feature.onboarding.ui.OnboardingScreen
+import com.jin.jjinweather.feature.onboarding.ui.OnboardingViewModel
+import com.jin.jjinweather.feature.temperature.ui.TemperatureScreen
+import com.jin.jjinweather.feature.temperature.ui.TemperatureViewModel
 import com.jin.jjinweather.ui.theme.JJinWeatherTheme
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
