+++ app/src/main/java/com/jin/jjinweather/feature/weather/domain/model/DailyWeather.kt
@@ -1,4 +1,4 @@
-package com.jin.jjinweather.feature.weather
+package com.jin.jjinweather.feature.weather.domain.model
 
 data class DailyWeather(
     val forecastDay: Long,
