+++ app/src/main/java/com/jin/jjinweather/feature/weatherimpl/data/WeatherDataSourceImpl.kt
@@ -5,9 +5,9 @@ import com.jin.jjinweather.BuildConfig
 import com.jin.jjinweather.feature.weather.data.OpenWeatherDataSource
 import com.jin.jjinweather.feature.weather.data.WeatherDataSource
 import com.jin.jjinweather.feature.weather.data.model.dto.WeatherDTO
-import com.jin.jjinweather.feature.weather.DailyWeather
-import com.jin.jjinweather.feature.weather.HourlyWeather
-import com.jin.jjinweather.feature.weather.Weather
+import com.jin.jjinweather.feature.weather.domain.model.DailyWeather
+import com.jin.jjinweather.feature.weather.domain.model.HourlyWeather
+import com.jin.jjinweather.feature.weather.domain.model.Weather
 import java.time.Instant
 
 class WeatherDataSourceImpl(
