--- app/src/main/java/com/jin/jjinweather/layer/data/datastore/DataStore.kt
@@ -1,8 +1,0 @@
-package com.jin.jjinweather.layer.data.datastore
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStore
-
-val Context.settingDataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
